Implement semaphore logic
Implemented the regular/exclusive semaphore based on the Linux
primitive semaphore (and mutex) objects.  Still need to determine
how to have the single netlock_release() function know how to
release either a LOCK_R or a LOCK_E.  This might need a
bookkeeping map.

Wait queues are not used in this implementation.  I'm not sure if
they're necessary at all, but they might make the bookkeeping
easier and allow us to avoid using the primitive semaphore and
mutex implementation.